+++ plugins/providers/virtualbox/action/import.rb
@@ -12,9 +12,6 @@ module VagrantPlugins
 
           # Import the virtual machine
           ovf_file = env[:machine].box.directory.join("box.ovf").to_s
-          if Vagrant::Util::Platform.cygwin?
-            ovf_file = `cygpath -m '#{ovf_file}'`.chomp
-          end
           env[:machine].id = env[:machine].provider.driver.import(ovf_file) do |progress|
             env[:ui].clear_line
             env[:ui].report_progress(progress, 100, false)