+++ plugins/providers/virtualbox/driver/base.rb
@@ -33,9 +33,6 @@ module VagrantPlugins
             if ENV.has_key?("VBOX_INSTALL_PATH")
               # Get the path.
               path = ENV["VBOX_INSTALL_PATH"]
-              if Vagrant::Util::Platform.cygwin?
-                path = `cygpath -u '#{path}'`.chomp
-              end
               @logger.debug("VBOX_INSTALL_PATH value: #{path}")
 
               # There can actually be multiple paths in here, so we need to
@@ -48,7 +45,7 @@ module VagrantPlugins
                 # and break out
                 vboxmanage = "#{path}VBoxManage.exe"
                 if File.file?(vboxmanage)
-                  @vboxmanage_path = vboxmanage
+                  @vboxmanage_path = Vagrant::Util::Platform.platform_path(vboxmanage)
                   break
                 end
               end